requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
awsebcli==3.14.11
billiard==3.5.0.4
blessed==1.15.0
boto3==1.9.1
botocore==1.12.98
cached-property==1.5.1
celery==4.2.1
cement==2.8.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
cryptography==2.3.1
cssselect==1.0.3
cssutils==1.0.2
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
dj-stripe==1.0.1
Django==2.0.7
django-admin-view-permission==1.9
django-appconf==1.0.2
django-autoslug-iplweb==1.9.4.dev0
django-braces==1.13.0
django-celery-email==2.0.1
django-cities-light==3.5.0
django-ckeditor==5.6.1
django-classy-tags==0.8.0
django-cms==3.5.3
django-compat==1.0.15
django-crispy-forms==1.7.0
django-formtools==2.1
django-google-maps==0.10.1
django-heroku==0.3.1
django-imagekit==4.0.2
django-jet==1.0.7
django-js-asset==1.1.0
django-modeltranslation==0.13b1
django-oauth-toolkit==1.2.0
django-polymorphic==2.0.3
django-rest-framework-social-oauth2==1.1.0
django-scheduler==0.8.8
django-sekizai==0.10.0
django-selectable==1.1.0
django-simple-mail==2.2.0
django-sns-view==0.1.2
django-social-auth==0.7.28
django-tables2==1.21.2
django-tagulous==0.13.2
django-threadlocals==0.8
django-tinymce==2.8.0
django-tinymce4-lite==1.7.4
django-treebeard==4.3
django-unique-slugify==1.1
django-vote==2.1.7
djangocms-admin-style==1.2.9
djangocms-installer==1.0.2
djangocms-text-ckeditor==3.7.0
djangorestframework==3.8.2
djappsettings==0.4.0
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
freezegun==0.3.10
future==0.16.0
gcloud==0.18.3
geocoder==1.38.1
google-api-python-client==1.4.1
googleapis-common-protos==1.5.3
gunicorn==19.9.0
html2text==2018.1.9
html5lib==1.0.1
httplib2==0.11.3
icalendar==4.0.2
idna==2.7
jmespath==0.9.3
jsmin==2.2.2
jsonfield==2.0.2
jsonschema==2.6.0
kombu==4.2.1
lxml==4.2.4
oauth2==1.9.0.post1
oauth2client==4.1.2
oauthlib==2.1.0
pathspec==0.5.5
pem==18.1.0
pilkit==2.0
Pillow==5.2.0
pluggy==0.7.1
premailer==3.2.0
progressbar2==3.39.2
protobuf==3.6.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python-doc-inherit==0.3.0
python-openid==2.2.5
python-social-auth==0.3.6
python-utils==2.3.0
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
ratelim==0.1.6
razorpay==1.1.1
requests==2.20.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rsa==3.4.2
s3transfer==0.1.13
semantic-version==2.5.0
shortuuid==0.5.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
stripe==2.8.1
termcolor==1.1.0
texttable==0.9.1
toml==0.9.6
tox==3.4.0
tqdm==4.26.0
tzlocal==1.5.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
var-dump==1.2
vine==1.1.4
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0
whitenoise==4.1.2
